2025/MSDEVMTL/2-Databricks-Chatbot-Companion/src/app/requirements.txt
python-dotenv
streamlit
streamlit-folium
folium
databricks-sdk